# api/docs/swagger.yaml
basePath: /
definitions:
  model.Post:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
      user_id:
        type: string
    required:
    - content
    - id
    - title
    - user_id
    type: object
  server.Error:
    properties:
      error: {}
    type: object
  server.Response:
    properties:
      data: {}
    type: object
host: localhost:8080
info:
  contact: {}
  description: api server for social network
  title: Social Network API
  version: "0.1"
paths:
  /api/v1/follow/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Target user ID to follow
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/server.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.Error'
      summary: Follow a user
      tags:
      - users
  /api/v1/posts/create:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Post'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/server.Error'
      summary: Create post
      tags:
      - posts
  /api/v1/unfollow/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Target user ID to Unfollow
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/server.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.Error'
      summary: Unfollow a user
      tags:
      - users
  /users/create:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "201":
          description: ok
          schema:
            type: string
      summary: Sign up
      tags:
      - users
swagger: "2.0"